tests/baseline/protocol-1.2/inspector/console/console-worker-nested-imports-syntax-error-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected identifier
Tests that nested import scripts in worker show correct stack on syntax error.

Message count: 1
invalidScript.js:1 Uncaught SyntaxError: Unexpected identifier
